+++ R/CLUSTERING-tsclust.R
@@ -403,10 +403,11 @@ tsclust <- function(series = NULL, type = "partitional", k = 2L, ...,
                 distmat <- dm$distmat
                 distmat_provided <- dm$distmat_provided
 
-                # Redefine new distmat
+                # Redefine new distmat if appropriate
                 control$distmat <- distmat
-                environment(family@dist)$control$distmat <- distmat
                 environment(family@allcent)$control$distmat <- distmat
+                if (!(distance == "dtw_lb" && !isTRUE(control$pam.precompute)))
+                    environment(family@dist)$control$distmat <- distmat
 
             } else {
                 distmat <- NULL